+++ src/config.rs
@@ -1,0 +1,10 @@
+pub const SCREEN_W: f32 = 800.0;
+pub const SCREEN_H: f32 = 600.0;
+
+pub const CRAB_H: f32 = 150.0;
+pub const CRAB_W: f32 = 100.0;
+pub const CRAB_S: f32 = 1.5;
+
+pub const CLAW_W: f32 = 14.0;
+pub const CLAW_H: f32 = 50.0;
+pub const CLAW_S: f32 = 5.0;
